+++ src/ui/react/settings/modules/developer/application.ts
@@ -24,6 +24,17 @@ export function setDevTools(devTools: boolean) {
   };
 }
 
+export function getUnlockShortcuts(): boolean {
+  return settings.value.unlockShortcuts;
+}
+
+export function setUnlockShortcuts(unlockShortcuts: boolean) {
+  settings.value = {
+    ...settings.value,
+    unlockShortcuts,
+  };
+}
+
 export function getUnstableOptimizations(): boolean {
   return settings.value.unstableOptimizations;
 }