+++ src/ui/react/settings/modules/developer/infra.tsx
@@ -1,7 +1,14 @@
 import { Switch, Tabs } from "antd";
 import { useTranslation } from "react-i18next";
 
-import { getDevTools, getUnstableOptimizations, setDevTools, setUnstableOptimizations } from "./application.ts";
+import {
+  getDevTools,
+  getUnlockShortcuts,
+  getUnstableOptimizations,
+  setDevTools,
+  setUnlockShortcuts,
+  setUnstableOptimizations,
+} from "./application.ts";
 
 import { SettingsGroup, SettingsOption } from "../../components/SettingsBox/index.tsx";
 import { ColorPalette } from "./ColorPalette.tsx";
@@ -12,6 +19,7 @@ import { WidgetsDebug } from "./WidgetsDebug.tsx";
 
 export function DeveloperTools() {
   const devTools = getDevTools();
+  const unlockShortcuts = getUnlockShortcuts();
   const unstableOptimizations = getUnstableOptimizations();
 
   const { t } = useTranslation();
@@ -19,19 +27,27 @@ export function DeveloperTools() {
   return (
     <>
       <SettingsGroup>
-        <SettingsOption>
-          <b>{t("devtools.enable")}</b>
-          <Switch value={devTools} onChange={setDevTools} />
-        </SettingsOption>
-      </SettingsGroup>
-
-      <SettingsGroup>
         <SettingsOption
-          label={t("devtools.unstable_optimizations")}
-          description={t("devtools.unstable_optimizations_description")}
-          action={<Switch value={unstableOptimizations} onChange={setUnstableOptimizations} />}
+          label={t("devtools.enable")}
+          description={t("devtools.enable_description")}
+          action={<Switch value={devTools} onChange={setDevTools} />}
         />
       </SettingsGroup>
+
+      {devTools && (
+        <SettingsGroup>
+          <SettingsOption
+            label={t("devtools.unstable_optimizations")}
+            description={t("devtools.unstable_optimizations_description")}
+            action={<Switch value={unstableOptimizations} onChange={setUnstableOptimizations} />}
+          />
+          <SettingsOption
+            label={t("devtools.unlock_shortcuts")}
+            description={t("devtools.unlock_shortcuts_description")}
+            action={<Switch value={unlockShortcuts} onChange={setUnlockShortcuts} />}
+          />
+        </SettingsGroup>
+      )}
 
       {devTools && (
         <Tabs