+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.7.1-dev]
+## [2.7.1]
 
 ### enhancements